requirements.txt
certifi
selenium
pandas
python-dotenv
chromedriver-autoinstaller
geckodriver-autoinstaller
urllib3
beautifulsoup4
